+++ src/lib/firebaseConfig.ts
@@ -1,0 +1,9 @@
+// Your web app's Firebase configuration
+export const firebaseConfig = {
+  apiKey: "AIzaSyDMdkst7F612rqEyu0iF86CaYOdJGBXPaw",
+  authDomain: "forgeite-frenzy.firebaseapp.com",
+  projectId: "forgeite-frenzy",
+  storageBucket: "forgeite-frenzy.appspot.com",
+  messagingSenderId: "928400685620",
+  appId: "1:928400685620:web:630f671a2fdaa4449b6861"
+};
